+++ packages/engine-canvas/index.js
@@ -10,7 +10,7 @@ export const run = (functionName, func, values, config) => {
   const mechanic = new Mechanic(func.settings, values, config);
 
   let isElAdded = false;
-  const onFrame = (el) => {
+  const onFrame = el => {
     if (!isElAdded) {
       isElAdded = true;
       root.appendChild(el);
@@ -20,6 +20,9 @@ export const run = (functionName, func, values, config) => {
     }
   };
   const onDone = async (el, name) => {
+    // TODO: This could go to core, because we always want to stop the drawloop
+    // when we're done drawing.
+    mechanic.drawLoop.stop();
     if (!isElAdded) {
       isElAdded = true;
       root.appendChild(el);
@@ -29,18 +32,21 @@ export const run = (functionName, func, values, config) => {
       mechanic.download(name || functionName);
     }
   };
-  const onSetState = async (obj) => {
+  const onSetState = async obj => {
     mechanic.setState(obj);
   };
+
+  const { drawLoop } = mechanic.callbacksForEngine();
 
   func.handler({
     inputs: mechanic.values,
     mechanic: {
       frame: onFrame,
       done: onDone,
+      drawLoop,
       state: mechanic.functionState,
-      setState: onSetState,
-    },
+      setState: onSetState
+    }
   });
   return mechanic;
 };